+++ packages/yourloops/lib/data/blip-api.ts
@@ -29,7 +29,7 @@
 import bows from "bows";
 
 import { PatientData } from "models/device-data";
-import { MessageNote } from "models/message";
+import MessageNote from "models/message";
 import { IUser } from "../../models/shoreline";
 import { User, AuthContext } from "../auth";
 import { t as translate } from "../language";